requirements.txt
Django==5.1.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
